:truck: Move PlainFileHandler into its own file
