+++ data_safe_haven/logging/logger.py
@@ -1,35 +1,13 @@
-"""Custom logging classes and functions to interact with Python logging"""
+"""Custom logging functions to interact with Python logging."""
 
 import logging
 from datetime import UTC, datetime
-from typing import Any
 
 from rich.logging import RichHandler
-from rich.text import Text
 
 from data_safe_haven.directories import log_dir
 
-
-class PlainFileHandler(logging.FileHandler):
-    """
-    Logging handler that cleans messages before sending them to a log file.
-    """
-
-    def __init__(self, *args: Any, **kwargs: Any):
-        """Constructor"""
-        super().__init__(*args, **kwargs)
-
-    @staticmethod
-    def strip_formatting(input_string: str) -> str:
-        """Strip console markup formatting from a string"""
-        text = Text.from_markup(input_string)
-        text.spans = []
-        return str(text)
-
-    def emit(self, record: logging.LogRecord) -> None:
-        """Emit a record without formatting"""
-        record.msg = self.strip_formatting(record.msg)
-        super().emit(record)
+from .plain_file_handler import PlainFileHandler
 
 
 def get_logger() -> logging.Logger:
